Support for 64 bit inode numbers. unfs3 now uses 64 bit wide variables to hold inode numbers. The file handle has also been extended. st_ino is 64 bit wide on 64 bit platforms and even on 32 bit platforms when using _FILE_OFFSET_BITS=64 (which we do, typically).
git-svn-id: svn://svn.code.sf.net/p/unfs3/code/unfs3.svn/trunk@331 fb7659fb-8760-4330-86f7-fdbeb49e81a3

+++ attr.c
@@ -184,7 +184,7 @@ post_op_attr get_post_buf(struct stat buf, struct svc_req * req)
 /*
  * lowlevel routine for getting post-operation attributes
  */
-static post_op_attr get_post_ll(const char *path, uint32 dev, uint32 ino,
+static post_op_attr get_post_ll(const char *path, uint32 dev, uint64 ino,
 				struct svc_req *req)
 {
     struct stat buf;
